+++ packages/happner-cluster/test/stress/users/child-client.js
@@ -1,0 +1,6 @@
+const testclient = require('../../_lib/client');
+try {
+testclient.create('user1', 'password', '6001');
+} catch (e) {
+    process.send(['dead', e])
+}