+++ packages/happner-client/RELEASES.md
@@ -298,6 +298,12 @@
 ----------------- 
   - TEN-34: updated happner-2
 
+12.0.4 2022-03-29
+-----------------
+  - TEN-92, TEN-93: logging levels and events
+  - TEN-102: loki snapshot  file redundancy
+  - TEN-92: made all dependencies test dependencies
+