+++ packages/happner-cluster/RELEASES.md
@@ -335,10 +335,15 @@
 ----------------- 
   - TEN-34: updated happner-client and happner-2
 
-12.0.5 2022-03-16
+12.0.5 2022-03-29
+-----------------
+  - TEN-92, TEN-93: logging levels and events
+  - TEN-102: loki snapshot  file redundancy
+  - TEN-92: made all dependencies test dependencies
+
+12.0.6 2022-03-16
 ----------------- 
   - SMC-4817: updated happn-cluster
-